refactor(core/utxo/starters): Added utxo starter and related changes to configure component scan
