--- components/core/src/main/java/com/bloxbean/cardano/yaci/store/YaciIndexerConfiguration.java
@@ -1,21 +1,0 @@
-package com.bloxbean.cardano.yaci.store;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@Configuration
-@EnableTransactionManagement
-@EnableScheduling
-@Slf4j
-public class YaciIndexerConfiguration {
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        return new ObjectMapper();
-    }
-
-}
