+++ components/core/src/main/java/com/bloxbean/cardano/yaci/store/core/util/StringUtil.java
@@ -1,0 +1,11 @@
+package com.bloxbean.cardano.yaci.store.core.util;
+
+public class StringUtil {
+
+    public static String nonNull(String input) {
+        if (input == null)
+            return "";
+        else
+            return input;
+    }
+}
