--- components/core/src/main/java/com/bloxbean/cardano/yaci/store/domain/Cursor.java
@@ -1,17 +1,0 @@
-package com.bloxbean.cardano.yaci.store.domain;
-
-import com.bloxbean.cardano.yaci.core.model.Era;
-import lombok.*;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@ToString
-@EqualsAndHashCode
-public class Cursor {
-    private Long slot;
-    private String blockHash;
-    private Long block;
-    private Era era;
-}
