--- components/core/src/main/java/com/bloxbean/cardano/yaci/store/util/StringUtil.java
@@ -1,11 +1,0 @@
-package com.bloxbean.cardano.yaci.store.util;
-
-public class StringUtil {
-
-    public static String nonNull(String input) {
-        if (input == null)
-            return "";
-        else
-            return input;
-    }
-}
