+++ settings.gradle
@@ -22,4 +22,5 @@ include 'components:stores:protocolparams'
 include 'components:stores:live'
 include 'components:submit'
 include ':starters:spring-boot-starter'
+include ':starters:utxo-spring-boot-starter'
