+++ starters/utxo-spring-boot-starter/src/main/java/com/bloxbean/cardano/yaci/store/starter/utxo/UtxoStoreAutoConfiguration.java
@@ -1,0 +1,18 @@
+package com.bloxbean.cardano.yaci.store.starter.utxo;
+
+import com.bloxbean.cardano.yaci.store.utxo.UtxoStoreConfiguration;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Import;
+
+@AutoConfiguration
+@EnableConfigurationProperties(UtxoStoreProperties.class)
+@Import(UtxoStoreConfiguration.class)
+@Slf4j
+public class UtxoStoreAutoConfiguration {
+
+    @Autowired
+    UtxoStoreProperties properties;
+}
