+++ starters/utxo-spring-boot-starter/src/test/java/com/bloxbean/cardano/yaci/store/starter/utxo/UtxoStoreAutoConfigurationTest.java
@@ -1,0 +1,5 @@
+package com.bloxbean.cardano.yaci.store.starter.utxo;
+
+class UtxoStoreAutoConfigurationTest {
+
+}
